+++ src/main/java/org/springframework/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -1,0 +1,31 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+/**
+ * 具备自动装配能力的BeanFactory
+ *
+ * @author Ling, Jiatong
+ * Date: 2022/11/21 14:48
+ */
+public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory{
+
+    @Override
+    protected Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException {
+        return doCreateBean(beanName, beanDefinition);
+    }
+
+
+    protected Object doCreateBean(String beanName, BeanDefinition beanDefinition) {
+        Class beanClass = beanDefinition.getBeanClass();
+        Object bean = null;
+        try {
+            bean = beanClass.newInstance();
+        } catch (Exception e) {
+            throw new BeansException("Instantiation of bean failed", e);
+        }
+        addSingleton(beanName, bean);
+        return bean;
+    }
+}