+++ src/main/java/org/springframework/beans/factory/support/AbstractBeanFactory.java
@@ -1,0 +1,29 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+/**
+ * BeanFactory实现
+ *
+ * @author Ling, Jiatong
+ * Date: 2022/11/21 14:20
+ */
+public abstract class AbstractBeanFactory extends DefaultSingletonBeanRegistry implements BeanFactory {
+
+    @Override
+    public Object getBean(String beanName) {
+        Object bean = getSingleton(beanName);
+        if (bean != null) {
+            return bean;
+        }
+
+        BeanDefinition beanDefinition = getBeanDefinition(beanName);
+        return createBean(beanName, beanDefinition);
+    }
+
+    protected abstract Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException;
+
+    protected abstract BeanDefinition getBeanDefinition(String beanName) throws BeansException;
+}