+++ src/main/java/org/springframework/beans/factory/support/DefaultListableBeanFactory.java
@@ -1,0 +1,32 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * 默认可列出的BeanFactory实现
+ *
+ * @author Ling, Jiatong
+ * Date: 2022/11/21 14:53
+ */
+public class DefaultListableBeanFactory extends AbstractAutowireCapableBeanFactory implements BeanDefinitionRegistry{
+
+    private Map<String, BeanDefinition> beanDefinitionMap = new HashMap<>();
+
+    @Override
+    protected BeanDefinition getBeanDefinition(String beanName) throws BeansException {
+        BeanDefinition beanDefinition = beanDefinitionMap.get(beanName);
+        if (beanDefinition == null) {
+            throw new BeansException("No bean named '" + beanName + "' is defined");
+        }
+        return beanDefinition;
+    }
+
+    @Override
+    public void registerBeanDefinition(String beanName, BeanDefinition beanDefinition) {
+        beanDefinitionMap.put(beanName, beanDefinition);
+    }
+}