+++ src/main/java/org/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -1,0 +1,26 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.factory.config.SingletonBeanRegistry;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * 默认的单例bean注册白噢实现类
+ *
+ * @author Ling, Jiatong
+ * Date: 2022/11/21 14:43
+ */
+public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
+
+    private Map<String, Object> singletonObjects = new HashMap<>();
+
+    @Override
+    public Object getSingleton(String beanName) {
+        return singletonObjects.get(beanName);
+    }
+
+    protected void addSingleton(String beanName, Object singletonObject) {
+        singletonObjects.put(beanName, singletonObject);
+    }
+}