+++ src/test/java/org/springframework/beans/factory/BeanDefinitionAndBeanDefinitionRegistryTest.java
@@ -1,0 +1,26 @@
+package org.springframework.beans.factory;
+
+import org.junit.Test;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.support.BeanDefinitionRegistry;
+import org.springframework.beans.factory.support.DefaultListableBeanFactory;
+
+/**
+ * @author Ling, Jiatong
+ * Date: 2022/11/21 15:38
+ */
+public class BeanDefinitionAndBeanDefinitionRegistryTest {
+
+
+    @Test
+    public void testBeanFactory() {
+        DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
+        BeanDefinition beanDefinition = new BeanDefinition(HelloService.class);
+        beanFactory.registerBeanDefinition("helloService", beanDefinition);
+
+
+        HelloService helloService = (HelloService) beanFactory.getBean("helloService");
+        helloService.sayHello();
+    }
+
+}