+++ changelog.md
@@ -38,5 +38,13 @@ public class SimpleBeanContainerTest {
 }
 ```
 
+## BeanDefinition和BeanDefinitionRegistry
 
+> 分支：bean-definition-and-bean-definition-registry
+
+主要增加如下类：
+
+* BeanDefinition，顾名思义，用于定义bean信息的类，包含bean的class类型、构造参数、属性值等信息，每个bean对应一个BeanDefinition的实例。简化BeanDefinition仅包含bean的class类型。
+* BeanDefinitionRegistry，BeaDefinition注册表接口，定义注册BeanDefinition的方法。
+* SingletonBeanRegistry及其实现类DefaultSingletonBeanRegistry，定义添加和获取单例bean的方法。